Move deferred token handling logic to DeferredToken class
